move the usecase tho schema file
