+++ server/internal/usecase/interactor/model.go
@@ -14,7 +14,6 @@ import (
 	"github.com/reearth/reearth-cms/server/pkg/model"
 	"github.com/reearth/reearth-cms/server/pkg/schema"
 	"github.com/reearth/reearth-cms/server/pkg/task"
-	"github.com/reearth/reearth-cms/server/pkg/value"
 	"github.com/reearth/reearthx/i18n"
 	"github.com/reearth/reearthx/log"
 	"github.com/reearth/reearthx/rerror"
@@ -176,156 +175,6 @@ func (i Model) Update(ctx context.Context, param interfaces.UpdateModelParam, op
 			}
 			return m, nil
 		})
-}
-
-func (i Model) UpdateWithNewSchemaFields(ctx context.Context, modelID id.ModelID, createFieldsParams []interfaces.CreateFieldParam, op *usecase.Operator) (*model.Model, error) {
-	return Run1(ctx, op, i.repos, Usecase().Transaction(),
-		func(ctx context.Context) (_ *model.Model, err error) {
-			if len(createFieldsParams) == 0 {
-				return nil, nil
-			}
-
-			m, err := i.repos.Model.FindByID(ctx, modelID)
-			if err != nil {
-				return nil, err
-			}
-
-			if !op.IsMaintainingProject(m.Project()) {
-				return nil, interfaces.ErrOperationDenied
-			}
-
-			s, err := i.repos.Schema.FindByID(ctx, m.Schema())
-			if err != nil {
-				return nil, err
-			}
-
-			// delete current fields if any
-			if len(s.Fields()) > 0 {
-				for _, field := range s.Fields() {
-					f := s.Field(field.ID())
-					if f == nil {
-						return nil, interfaces.ErrFieldNotFound
-					}
-					if f.Type() == value.TypeReference {
-						err := i.deleteCorrespondingField(ctx, s, f)
-						if err != nil {
-							return nil, err
-						}
-					}
-					s.RemoveField(field.ID())
-				}
-			}
-
-			// create new fields
-			for _, createFieldParam := range createFieldsParams {
-				if createFieldParam.Key == "" || s.HasFieldByKey(createFieldParam.Key) {
-					return nil, schema.ErrInvalidKey
-				}
-
-				newField, err := schema.NewField(createFieldParam.TypeProperty).
-					NewID().
-					Unique(createFieldParam.Unique).
-					Multiple(createFieldParam.Multiple).
-					Required(createFieldParam.Required).
-					Name(createFieldParam.Name).
-					Description(lo.FromPtr(createFieldParam.Description)).
-					Key(id.NewKey(createFieldParam.Key)).
-					DefaultValue(createFieldParam.DefaultValue).
-					Build()
-				if err != nil {
-					return nil, err
-				}
-
-				if createFieldParam.Type == value.TypeReference {
-					err = i.createCorrespondingField(ctx, s, newField, createFieldParam)
-					if err != nil {
-						return nil, err
-					}
-				}
-
-				if createFieldParam.Type == value.TypeGroup {
-					var g *schema.FieldGroup
-					createFieldParam.TypeProperty.Match(schema.TypePropertyMatch{
-						Group: func(f *schema.FieldGroup) {
-							g = f
-						},
-					})
-					_, err = i.repos.Group.FindByID(ctx, g.Group())
-					if err != nil {
-						return nil, err
-					}
-				}
-
-				s.AddField(newField)
-
-				if err := setTitleField(&createFieldParam.IsTitle, s, newField.ID().Ref()); err != nil {
-					return nil, err
-				}
-			}
-
-			if err := i.repos.Schema.Save(ctx, s); err != nil {
-				return nil, err
-			}
-
-			return m, nil
-		})
-}
-
-func (i Model) deleteCorrespondingField(ctx context.Context, s *schema.Schema, f *schema.Field) error {
-	fr, _ := schema.FieldReferenceFromTypeProperty(f.TypeProperty())
-	if fr.CorrespondingFieldID() == nil {
-		return nil
-	}
-
-	rs := s
-	if s.ID() != fr.Schema() {
-		s, err := i.repos.Schema.FindByID(ctx, fr.Schema())
-		if err != nil {
-			return err
-		}
-		rs = s
-	}
-
-	rs.RemoveField(*fr.CorrespondingFieldID())
-	if err := i.repos.Schema.Save(ctx, rs); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (i Model) createCorrespondingField(ctx context.Context, s *schema.Schema, f *schema.Field, param interfaces.CreateFieldParam) error {
-	rInput, _ := schema.FieldReferenceFromTypeProperty(param.TypeProperty)
-	// if the corresponding field is not passed it's not two-way
-	if rInput.CorrespondingField() == nil {
-		return nil
-	}
-
-	rs := s
-	if s.ID() != rInput.Schema() {
-		s, err := i.repos.Schema.FindByID(ctx, rInput.Schema())
-		if err != nil {
-			return err
-		}
-		rs = s
-	}
-
-	if rs.HasFieldByKey(rInput.CorrespondingField().Key) {
-		return interfaces.ErrReferencedFiledKeyExists
-	}
-
-	cf, err := schema.CreateCorrespondingField(s.ID(), *param.ModelID, f, *rInput.CorrespondingField())
-	if err != nil {
-		return err
-	}
-
-	rs.AddField(cf)
-
-	if err := i.repos.Schema.Save(ctx, rs); err != nil {
-		return err
-	}
-
-	return nil
 }
 
 func (i Model) CheckKey(ctx context.Context, pId id.ProjectID, s string) (bool, error) {
