+++ server/internal/usecase/interfaces/model.go
@@ -71,7 +71,6 @@ type Model interface {
 	Create(context.Context, CreateModelParam, *usecase.Operator) (*model.Model, error)
 	Update(context.Context, UpdateModelParam, *usecase.Operator) (*model.Model, error)
 	UpdateOrder(context.Context, id.ModelIDList, *usecase.Operator) (model.List, error)
-	UpdateWithNewSchemaFields(context.Context, id.ModelID, []CreateFieldParam, *usecase.Operator) (*model.Model, error)
 	CheckKey(context.Context, id.ProjectID, string) (bool, error)
 	Delete(context.Context, id.ModelID, *usecase.Operator) error
 	Publish(context.Context, []PublishModelParam, *usecase.Operator) error
