+++ server/internal/usecase/interfaces/schema.go
@@ -42,6 +42,12 @@ type UpdateFieldParam struct {
 	DefaultValue *value.Multiple
 }
 
+type ModelData struct {
+	ModelID   *id.ModelID
+	SchemaID  id.SchemaID
+	ProjectID id.ProjectID
+}
+
 var (
 	ErrInvalidTypeProperty       = rerror.NewE(i18n.T("invalid type property"))
 	ErrReferencedFiledKeyExists  = rerror.NewE(i18n.T("referenced field key exists"))
@@ -57,6 +63,7 @@ type Schema interface {
 	FindByIDs(context.Context, []id.SchemaID, *usecase.Operator) (schema.List, error)
 	FindByModel(context.Context, id.ModelID, *usecase.Operator) (*schema.Package, error)
 	CreateField(context.Context, CreateFieldParam, *usecase.Operator) (*schema.Field, error)
+	CreateFieldsForModel(context.Context, ModelData, []CreateFieldParam, *usecase.Operator) (*schema.Schema, error)
 	UpdateField(context.Context, UpdateFieldParam, *usecase.Operator) (*schema.Field, error)
 	UpdateFields(context.Context, id.SchemaID, []UpdateFieldParam, *usecase.Operator) (schema.FieldList, error)
 	DeleteField(context.Context, id.SchemaID, id.FieldID, *usecase.Operator) error
